requirements.txt
anyio==4.0.0
black==23.9.1
certifi==2023.7.22
click==8.1.7
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.10.0
PyYAML==6.0.1
sniffio==1.3.0
